Revert pull request #161 to fix freezing-page problem
Pull request #161 was the one introducing a second visualization to display weighted percentage of known words. The reason for undoing the pull request is that the changes introduced seem to be the cause of the browser-page-freeze that happens when you click the vocab list dropdown and select a vocab list. The reason for the freeze is likely that the changes introduced in the pull request are a bit heavy to run on the front-end. I noticed that when you open the Chrome Task Manager and select a vocab list, the CPU usage spikes to 100+ and stays as such indefinitely. Thus, perhaps the computation introduced in the PR is heavy for running on the front-end or there is a bug in the code change in that PR. The real reason is yet TBD. At any rate, reverting the pull request resolves the problem.