Fix django debug
Shell environment variables are strings. As such, when you pass in 'False' into the 'DJANGO_DEBUG' env var, it is parsed as a True-ey string in settings.py, which is not what we want. Instead of 'False', set DJANGO_DEBUG env var to 0. It is still a string, but it is Falsy.