Fix up personal vocab list view
Made changes to vocab list entry models to make sure that the link function always runs when a new entry is created. Views and other model consumers no longer need to run `link_job()` on a newly-created vocab list entry, since that's been added to the save method on the abstract class.

On the front end, this makes sure that the new "definition" properties are used instead of "gloss" properties. The export function has been updated to export definitions properly, and to export lemma_ids as well.