Make the vocab list upload better
Old behavior: A vocab upload list needed to be a two column tsv with headword in the first column and definition in the second column. Any other columns would break the upload, and headers would be interpreted as data. No other information could be imported.

New behavior: Vocab list upload files are still expected to be well-formatted TSVs, but now they should have column headers that match the fields in a vocab list entry model. Any field on the model can be a column header in the tsv, although foreign key relationships, like the relationship to lemma, need to have "_id" appended. Any invalid columns will be silently ignored.

The export has also been changed to match this format, such that if you export a vocab list, the exported tsv should be able to be imported into a new vocab list with the same lemma connection and familiarity rating, if applicable.

Last but not least, the import page has been modified to include a downloadable template file.