fix quick add
This commit fixes the quick add component to work without the lattice. Quick add prompts users for a headword and definition, and as they type the headword, a select element fills with lemmas to connect to. The selected lemma is linked to the vocab list entry when it is added to the list.

This commit also makes some changes to the api and models in django. This improves error logging on the quick add endpoint and fixes an error that would occur when the `data` method was called on vocab list entries that didn't have an associated lemma.