switched arrows and added default link
I think we had left left-pointing arrows in while testing to see if we were editing the right files, so I removed those and added a default case for the link in the case that the http referer is empty. If there's a direct link to the page from elsewhere, like if someone links to a text from Canvas, then the referer is empty and the breadcrumb goes nowhere. I changed it so if there's no http referer, it gets the url for the lemmatized texts list view, as named in urls.py under lemmatized texts.